MESSAGING SYSTEM WORKS
So, clients can write messages and they get added to the messages
waiting for world list. We just need to make sure that the game can get
those messages and do the right things with them.

Oh yeah, clients can't receive messages from the world yet. Which is a
bummer. Working on that.